ios: remove unused network discovery
